% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junyuan_li.R
\name{table_bind_group}
\alias{table_bind_group}
\title{Combine groups by column}
\usage{
table_bind_group(
  object,
  id,
  col,
  label = "groups",
  sep = ";",
  keep = FALSE,
  unique = TRUE,
  count = FALSE,
  label_count = "count"
)
}
\arguments{
\item{object}{a data frame with two or more columns}

\item{id}{the column for group}

\item{col}{the coloumn for combinning}

\item{label}{the new column name}

\item{sep}{the separator between combined group}

\item{keep}{keep other columns}

\item{unique}{unique while combinning}

\item{count}{count groups}

\item{label_count}{the new count column name}
}
\description{
Combine groups by column
}
