% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ziyu_zhang.R
\name{table_correct_batch}
\alias{table_correct_batch}
\title{Batch effect correction}
\usage{
table_correct_batch(object, method, plot, meta_data, id = NULL, batch, group)
}
\arguments{
\item{object}{a data frame}

\item{method}{raw, combat, limma, mnn, harmony, cca}

\item{plot}{true, false}

\item{meta_data}{meta data with batch aand group}

\item{id}{the id column for future development}

\item{batch}{the batch to correct}

\item{group}{the meaningful group}
}
\description{
Batch effect correction
}
