% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junyuan_li.R
\name{table_split_group}
\alias{table_split_group}
\title{Split groups by column}
\usage{
table_split_group(object, col, header = "", sep = "")
}
\arguments{
\item{object}{a data frame with two or more columns}

\item{col}{the coloumn for split}

\item{header}{header string to ignore}

\item{sep}{the split separator in string}
}
\description{
Split groups by column
}
